loaders/vector/requirements.txt
# Aerospike dependencies
aerospike-vector-search
aerospike

# Vector dependencies
sentence_transformers
torch
Pillow